Add run-workdir-check.sh
Since the workdir can now be specified using an argument to the
entrypoint or as an argument to docker, we need to test that the workdir
is as expected in the container for either case.

run-workdir-check.sh allows us to do the check for both when runtests.sh
sees that run-workdir-check.sh is the test being executed.


squash

+++ tests/runtests.sh
@@ -26,10 +26,23 @@ for i in run-*.sh; do
     # get absolute path to it
     LOCAL_WDIR=$(readlink -f ${LOCAL_WDIR})
     cp $i $LOCAL_WDIR
-    docker run --rm -t -v $LOCAL_WDIR:/workdir \
-	--workdir /workdir \
-	$IMAGE \
-	/workdir/$i
+    if [ "$i" == "run-workdir-check.sh" ]; then
+	# This is to make sure the workdir is correct when using either docker
+	# or the entrypoint workdir argument.
+	docker run --rm -t -v $LOCAL_WDIR:/workdir \
+	    --workdir /workdir \
+	    $IMAGE \
+	    /workdir/$i &&
+	docker run --rm -t -v $LOCAL_WDIR:/workdir \
+	    $IMAGE \
+	    --workdir /workdir \
+	    /workdir/$i
+    else
+	docker run --rm -t -v $LOCAL_WDIR:/workdir \
+	    $IMAGE \
+	    --workdir /workdir \
+	    /workdir/$i
+    fi
     RET=$?
     if [ ${RET} != 0 ]; then
 	echo "Test $i failed"
